Refactor command system into more files and make it more modular
